Module Terminal.Size

val sigwinch : int option

The number of the signal used to indicate terminal size changes. None on Windows.

Functions for getting the size of the terminal to which stdout is attached (provided stdout is a TTY).

type dimensions = {

rows : int;

columns : int;

}

val get_dimensions : unit -> dimensions option

val get_columns : unit -> int option

val get_rows : unit -> int option