Module Terminal.Style

type t = Terminal_ansi.Style.t

The type of terminal styles: values that can be printed to a terminal in order to change the way that it renders text.

val code : t -> string

Get the ANSI escape code for the given style.

Constructing ANSI styles:

val none : t

val bold : t

val faint : t

val italic : t

val underline : t

val reverse : t

val fg : Color.t -> t

val bg : Color.t -> t